Added a test for toString all the values
